+++ src/otx/algo/detection/losses/__init__.py
@@ -1,9 +1,10 @@
-# Copyright (C) 2024 Intel Corporation
+# Copyright (C) 2024-2025 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 #
 """Custom OTX Losses for Object Detection."""
 
 from .atss_loss import ATSSCriterion
+from .dfine_loss import DFINECriterion
 from .rtdetr_loss import DetrCriterion
 from .rtmdet_loss import RTMDetCriterion
 from .ssd_loss import SSDCriterion
@@ -17,4 +18,5 @@ __all__ = [
     "SSDCriterion",
     "YOLOv9Criterion",
     "YOLOXCriterion",
+    "DFINECriterion",
 ]