+++ tests/integration/api/test_xai.py
@@ -51,6 +51,11 @@ def test_forward_explain(
         # TODO(Eugene): maskdino not support yet.
         pytest.skip(f"There's issue with inst-seg: {model_name}. Skip for now.")
 
+    if "dfine" in model_name:
+        # TODO(Eugene): dfine not support yet.
+        # https://jira.devtools.intel.com/browse/CVS-160781
+        pytest.skip(f"There's issue with dfine: {model_name}. Skip for now.")
+
     if "dino" in model_name:
         pytest.skip("DINO is not supported.")
 
@@ -117,6 +122,11 @@ def test_predict_with_explain(
     if "maskdino" in model_name:
         # TODO(Eugene): maskdino not support yet.
         pytest.skip(f"There's issue with inst-seg: {model_name}. Skip for now.")
+
+    if "dfine" in model_name:
+        # TODO(Eugene): dfine not support yet.
+        # https://jira.devtools.intel.com/browse/CVS-160781
+        pytest.skip(f"There's issue with dfine: {model_name}. Skip for now.")
 
     if "rtmdet_tiny" in recipe:
         # TODO (sungchul): enable xai for rtmdet_tiny (CVS-142651)